% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_anno.R
\name{format_gene_go_tooltip_batch}
\alias{format_gene_go_tooltip_batch}
\title{Format GO annotaion for each gene}
\usage{
format_gene_go_tooltip_batch(
  genes,
  orgdb_pkg = "org.Hs.eg.db",
  keytype = "SYMBOL",
  n_each = 2
)
}
\arguments{
\item{genes}{genes to annotation.}

\item{orgdb_pkg}{package for GO annotation.}

\item{keytype}{type for gene name.}

\item{n_each}{top n item to show.}
}
\description{
Format GO annotaion for each gene
}
