% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-ivolcano-line.R
\name{geom_ivolcano_line}
\alias{geom_ivolcano_line}
\title{geom_ivolcano_line}
\usage{
geom_ivolcano_line(linetype = "longdash", color = "grey40", linewidth = 0.5)
}
\arguments{
\item{linetype}{line type for the threshold lines}

\item{color}{line color for the threshold lines}

\item{linewidth}{line width for the threshold lines}
}
\value{
A ggplot2 layer object
}
\description{
Add threshold lines to an iVolcano plot
}
