% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_anno.R
\name{get_gene_go_summary_batch}
\alias{get_gene_go_summary_batch}
\title{Get GO annotaion for each gene}
\usage{
get_gene_go_summary_batch(
  genes,
  orgdb_pkg = "org.Hs.eg.db",
  keytype = "SYMBOL",
  n_each = 2
)
}
\arguments{
\item{genes}{genes to annotation.}

\item{orgdb_pkg}{package for GO annotation.}

\item{keytype}{type for gene name.}

\item{n_each}{top n item to show.}
}
\description{
Get GO annotaion for each gene
}
