% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idotplot.R
\name{idotplot}
\alias{idotplot}
\title{idotplot}
\usage{
idotplot(
  object,
  x = "GeneRatio",
  color = "p.adjust",
  showCategory = 10,
  size = "Count",
  label_format = 30,
  trigger = c("click", "hover"),
  title = "",
  ...
)
}
\arguments{
\item{object}{enrichment result object (e.g. enrichResult) or a data frame}

\item{x}{x-axis variable, one of 'GeneRatio', 'Count', etc.}

\item{color}{variable for color, e.g., 'p.adjust'}

\item{showCategory}{number of categories to show}

\item{size}{variable for size, e.g., 'Count'}

\item{label_format}{a numeric value sets wrap length, alternatively a custom function to format axis labels.}

\item{trigger}{interaction trigger, one of "click" or "hover"}

\item{title}{plot title}

\item{...}{additional parameters passed to fortify}
}
\value{
ggplot object with interactive layers
}
\description{
Interactive Dot Plot
}
\author{
Guangchuang Yu
}
