% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-ivolcano-gene.R, R/ivolcano.R
\name{geom_ivolcano_gene}
\alias{geom_ivolcano_gene}
\alias{ivolcano}
\title{geom_ivolcano_gene}
\usage{
geom_ivolcano_gene(
  top_n = 10,
  label_mode = "separate",
  fontface = "italic",
  label_sig_only = TRUE,
  filter = NULL,
  orgdb_pkg = NULL,
  orgdb_keytype = "SYMBOL",
  go_n_each = 2
)

ivolcano(
  data,
  logFC_col = "logFC",
  pval_col = "adj.P.Val",
  gene_col = "gene",
  title = "Volcano Plot",
  interactive = TRUE,
  onclick_fun = NULL,
  pval_cutoff = 0.05,
  logFC_cutoff = 1,
  pval_cutoff2 = NULL,
  logFC_cutoff2 = NULL,
  size_by = "none",
  point_size = list(base = 2, medium = 4, large = 6),
  threshold_line = list(color = "black", linetype = "dashed", linewidth = 0.5),
  top_n = 10,
  label_mode = "separate",
  fontface = "italic",
  label_sig_only = TRUE,
  filter = NULL,
  orgdb_pkg = NULL,
  orgdb_keytype = "SYMBOL",
  go_n_each = 2
)
}
\arguments{
\item{top_n}{top N genes to display the labels (gene ID)}

\item{label_mode}{one of 'all' or 'separate' (default).
If label_mode = 'all', top_n genes with minimal p values will be displayed,
otherwise, top_n up-regulated and top_n down-regulated genes will be displayed.}

\item{fontface}{one of 'plain', 'bold', 'italic' (default) and their combination, e.g. 'bold.italic'}

\item{label_sig_only}{whether filter significant genes before subset 'top_n' genes}

\item{filter}{custom filter expression to select genes for labeling}

\item{orgdb_pkg}{package for GO annotation.}

\item{orgdb_keytype}{type for gene name.}

\item{go_n_each}{top n GO item to show.}

\item{data}{A data frame that contains minimal information with gene id, logFC and adjusted P values}

\item{logFC_col}{column name in 'data' that stored the logFC values}

\item{pval_col}{column name in 'data' that stored the adjusted P values}

\item{gene_col}{column name in 'data' that stored the gene IDs}

\item{title}{plot title}

\item{interactive}{whether plot the graph in interactive mode}

\item{onclick_fun}{effects when click on the dot (gene), default is NULL}

\item{pval_cutoff}{cutoff of the adjusted P values}

\item{logFC_cutoff}{cutoff of the logFC values}

\item{pval_cutoff2}{second cutoff of the adjusted P values for advanced mode}

\item{logFC_cutoff2}{second cutoff of the logFC values for advanced mode}

\item{size_by}{one of "none" (default), "manual" (set by \code{point_size}), "negLogP", "absLogFC", or other variable in the input data to scale dot sizes.}

\item{point_size}{set point size when \code{size_by} is "manual", a list with three elements: base, medium, large.}

\item{threshold_line}{customize threshold line style (e.g., line color, type, and width)}
}
\value{
volcano plot
}
\description{
Add gene labels to ivolcano plot

interactive volcano plot
}
\examples{
# example data
f <- system.file("extdata/airway.rds", package = "ivolcano")
df <- readRDS(f)
# plot
ivolcano(df,
  logFC_col = "log2FoldChange",
  pval_col = "padj",
  gene_col = "symbol",
  onclick_fun = onclick_genecards
)
}
\author{
Guangchuang Yu
}
