% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick_ensembl}
\alias{onclick_ensembl}
\title{onclick_ensembl}
\usage{
onclick_ensembl(gene)
}
\arguments{
\item{gene}{query gene}
}
\value{
onclick action
}
\description{
onclick function to popup Ensembl gene webpage
}
\author{
Guangchuang Yu
}
