% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick_fanyi}
\alias{onclick_fanyi}
\title{onclick_fanyi}
\usage{
onclick_fanyi(gene_summary, cols)
}
\arguments{
\item{gene_summary}{output from \code{fanyi::gene_summary()}}

\item{cols}{selected columns from 'gene_summary'}
}
\value{
onclick function
}
\description{
generate onclick function from \code{fanyi::gene_summary()} output
}
\author{
Guangchuang Yu
}
