% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick_genecards}
\alias{onclick_genecards}
\title{onclick_genecards}
\usage{
onclick_genecards(gene)
}
\arguments{
\item{gene}{query gene}
}
\value{
onclick action
}
\description{
onclick function to popup genecards webpage
}
\author{
Guangchuang Yu
}
