% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick_hgnc}
\alias{onclick_hgnc}
\title{onclick_hgnc}
\usage{
onclick_hgnc(gene)
}
\arguments{
\item{gene}{query gene}
}
\value{
onclick action
}
\description{
onclick function to popup HGNC gene webpage
}
\author{
Guangchuang Yu
}
