% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick_ncbi}
\alias{onclick_ncbi}
\title{onclick_ncbi}
\usage{
onclick_ncbi(gene)
}
\arguments{
\item{gene}{query gene}
}
\value{
onclick action
}
\description{
onclick function to popup NCBI gene webpage
}
\author{
Guangchuang Yu
}
