% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick_pubmed}
\alias{onclick_pubmed}
\title{onclick_pubmed}
\usage{
onclick_pubmed(gene)
}
\arguments{
\item{gene}{query gene}
}
\value{
onclick action
}
\description{
onclick function to popup PubMed webpage
}
\author{
Guangchuang Yu
}
