% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick_uniprot}
\alias{onclick_uniprot}
\title{onclick_uniprot}
\usage{
onclick_uniprot(gene)
}
\arguments{
\item{gene}{query gene}
}
\value{
onclick action
}
\description{
onclick function to popup UniProt webpage
}
\author{
Guangchuang Yu
}
