% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_volcano.R
\name{pathway_volcano}
\alias{pathway_volcano}
\title{pathway_volcano}
\usage{
pathway_volcano(
  p1,
  p2,
  widths = c(1, 1),
  trigger = c("click", "hover"),
  hover_css = "stroke:red;stroke-width:2px;",
  hover_inv_css = "opacity:0.15;",
  tooltip_css =
    "background:#fff;color:#000;border:1px solid #999;padding:6px;border-radius:4px;",
  non_selected_color = "#BDBDBD",
  non_selected_opacity = 0.25,
  label_follow = TRUE,
  ...
)
}
\arguments{
\item{p1}{interactive dot plot (from idotplot)}

\item{p2}{interactive volcano plot (from ivolcano)}

\item{widths}{relative widths of the two plots, default is c(1, 1)}

\item{trigger}{interaction trigger, one of "click" or "hover"}

\item{hover_css}{css for hovered elements}

\item{hover_inv_css}{css for non-hovered elements}

\item{tooltip_css}{css for tooltip}

\item{non_selected_color}{color for non-selected volcano points}

\item{non_selected_opacity}{opacity for non-selected volcano points}

\item{label_follow}{whether volcano labels follow pathway selection}

\item{...}{additional parameters passed to ggiraph::girafe}
}
\value{
a girafe object
}
\description{
Combine Interactive Pathway Dot Plot and Volcano Plot
}
\author{
Guangchuang Yu
}
