% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_figureya}
\alias{scale_color_figureya}
\title{scale_color_figureya}
\usage{
scale_color_figureya(
  mode = c("classic", "advanced"),
  na.value = "grey80",
  guide = "none",
  ...
)
}
\arguments{
\item{mode}{one of "classic" or "advanced"}

\item{na.value}{color for NA values}

\item{guide}{guide parameter passed to ggplot2::scale_color_manual}

\item{...}{additional parameters passed to ggplot2::scale_color_manual}
}
\value{
A ggplot2 scale object
}
\description{
scale_color_figureya
}
